+++ src/index.js
@@ -2,6 +2,8 @@ import './editors/wrapper-editor.js';
 import './cards/wrapper-card.js';
 import './cards/title-card.js';
 import './cards/light-card.js';
+import './cards/switch-card.js';
+import './cards/sensor-card.js';
 import './cards/shutter-card.js';
 import './cards/navigation-card.js';